requirements.txt
# --- Interface ---
streamlit==1.33.0

# --- LangChain Framework ---
langchain==0.2.1
langchain-openai==0.1.7       # ✅ Matche bien langchain==0.2.1

# --- Vector Stores ---
chromadb==0.5.3
faiss-cpu==1.13.0

# --- Connecteurs OpenAI ---
openai==1.30.1
tiktoken==0.7.0

# --- Utilitaires ---
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pandas==2.2.2
requests==2.31.0
pydantic==2.7.1
numpy==1.26.4